changed gps_fix to uint16_t in flight packet

+++ flight_packet.h
@@ -6,7 +6,7 @@
 struct FlightPacket {
     uint32_t timestamp_utc;     // 4 bytes
     uint8_t fsm_mode;           // 1
-    uint8_t gps_fix;            // 1
+    uint16_t gps_fix;           // 2
     float heading_deg;          // 4
     float target_heading_deg;   // 4
     float h_speed_m_s;      // 4
